Added download count (total and per version), and changed yearly commit count
Changed the yearly commit count by splitting it into 2 different functions. One takes a year as input and gives the number of commits within that year. The other one return the number of commits within the last year (starting from the current week).

Added 2 functions to get the number of downloads. One of these returns the number of downloads of the repo in total, and one returns the number of downloads of a specific release.

Also updated the way we checked if there were multiple result pages. Instead of just seeing if there were links in the link header, we now check if there is a 'next' or 'last' link. This is important, as the link header might contain a 'prev' link, which does indicate multiple pages, but not that we are on the last one.