Refactoring, interface making and PEP8 conformity
Refactored github_api_calls and github_get_token so that it only gets the API when needed. Also edited some names for clarity.

Made a basic interface file which can be used to get the wanted data-points without having to worry what goes on "under the hood".
Currently, only the API call functions are present

Switched from Pylance to Flake8 for PEP8 checking