Started the testing process
The general testing is done using Pytest.
All the web requests are handled using Responses, which allows you to set what a request should return.

The testing files are currently mixed with the other files. This is only due to the fact that imports would not work if the files were anywhere else. Will look into a fix for this in the future