Finished the tests for the API call itself
There are now tests in place for all of the functions having to do with the API calling itself. Functions that request data through an API call are not tested yet.

I did however add some more error-checking in those functions, and will add tests for them later

I also updated the demo so that you can specify that you also want to see the search results. The README has been updated to reflect this