Refactored the spider, finished the tests
I have refactored the spider so that it can get the open and closed issue counts separately, instead of only the open issue count. With this I have also refactored the issue ratio function to now use the separate open and closed issue functions, instead of finding the closed issue count itself.

I also added the final tests for the spider