Fixed bug in the tests for the API calls
I had forgotten that the .env file might not exist. This caused the program to try and authenticate the user, but as this requires user input, the tests got stuck waiting.

Now, I made a separate test for when the key does not exists, and made sure the program can run without an .env file