Updated controller and started on tests
Changed the Controller class to separate the virus data-point grabbing from the run() function by giving it it's own function (get_virus_data).

Added a base testing file for the controller unit tests, and implemented the virus scanning test. This should suffice as a template for how the other functions should be tested.